set all IteratorsConfig fields to u32
Surely that should be more than enough.

+++ src/redisearch_rs/rqe_iterators/src/config.rs
@@ -9,23 +9,21 @@
 
 //! Configuration parameters related to query execution.
 
-use std::ffi::{c_longlong, c_uint};
-
 /// Configuration parameters related to the query execution.
 #[cheadergen::config(export, rename_all = "camelCase")]
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 #[repr(C)]
 pub struct IteratorsConfig {
     /// The maximal number of expansions we allow for a prefix. Default: 200
-    pub max_prefix_expansions: c_longlong,
+    pub max_prefix_expansions: u32,
     /// The minimal number of characters we allow expansion for in a prefix
     /// search. Default: 2
-    pub min_term_prefix: c_longlong,
+    pub min_term_prefix: u32,
     /// The minimal word length to stem. Default: 4
-    pub min_stem_length: c_uint,
+    pub min_stem_length: u32,
     /// Minimum number of children for a union iterator to use a heap-based
     /// implementation instead of a flat (linear scan) one. Default: 20
-    pub min_union_iter_heap: c_longlong,
+    pub min_union_iter_heap: u32,
 }
 
 impl Default for IteratorsConfig {
